+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT VERSION
 
 - refactor: remove dependent on `react-draggable`
+- fix: undefined parentId should be considered as root item
 
 ## v1.4.0 (2019-07-01)