+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - refactor: remove dependent on `react-draggable`
 - fix: undefined parentId should be considered as root item
+- fix: `flattenOnKeys` not works with immutable data (regression from #23)
 
 ## v1.4.0 (2019-07-01)